bigint: Split `PrivateExponent` into its own submodule.
Better encapsulate `PrivateExponent` and its primary use, the constant-time
exponentiation.

`git diff HEAD^1:src/arithmetic/bigint.rs src/arithmetic/bigint/private_exponent.rs`
